Create klocktal-pull.php
This can be integrated into a web solution, but I currently just run it from the command line.  It connect to one of my zcash accounts and reads a code 10402 (it actually just skips over this, but it will eventually be a description of what is to follow.  Then it reads the next two transactions as four digit codes, outputting the kloctal message 'dog'.